address comments (refactor package names)
